+++ threestudio/models/renderers/__init__.py
@@ -4,4 +4,5 @@ from . import (
     nerf_volume_renderer,
     neus_volume_renderer,
     nvdiff_rasterizer,
+    patch_renderer
 )